fix(refactor): some refactoring for mock es data
